# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SVC Ecommerce API",
        "description": "API documentation for SVC Coffee Ecommerce Platform - Complete e-commerce solution with cart, checkout, and order management",
        "contact": {
            "name": "SVC Development Team",
            "email": "usman@workforcecommerce.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://workforcecommerce.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://staging.workforcecommerce.com",
            "description": "Staging Server"
        },
        {
            "url": "https://api.workforcecommerce.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/orders/{orderNumber}/return": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Request return/refund",
                "description": "Request return or refund for order items",
                "operationId": "requestReturn",
                "parameters": [
                    {
                        "name": "orderNumber",
                        "in": "path",
                        "description": "Order number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ORD-2025-00001"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items",
                                    "return_method"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "item_id": {
                                                    "description": "Order item UUID",
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "minimum": 1,
                                                    "example": 1
                                                },
                                                "reason": {
                                                    "type": "string",
                                                    "example": "Product damaged during shipping"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "return_method": {
                                        "type": "string",
                                        "enum": [
                                            "refund",
                                            "replacement"
                                        ],
                                        "example": "refund"
                                    },
                                    "additional_notes": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "example": "Please process refund to original payment method",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return request submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Return request submitted successfully. Our team will review and contact you soon."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Order is not eligible for return/refund",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order is not eligible for return/refund"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Enter your API key",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}